duplication validation added for vendors and manufacturer
