+++ src/Presentation/Nop.Web/Administration/Validators/Catalog/CategoryValidator.cs
@@ -1,5 +1,4 @@
-﻿using System.Linq;
-using FluentValidation;
+﻿using FluentValidation;
 using Nop.Admin.Models.Catalog;
 using Nop.Services.Localization;
 using Nop.Web.Framework.Validators;
@@ -12,17 +11,7 @@ namespace Nop.Admin.Validators.Catalog
         {
             RuleFor(x => x.Name).NotEmpty().WithMessage(localizationService.GetResource("Admin.Catalog.Categories.Fields.Name.Required"));
 
-            RuleFor(x => x.PageSizeOptions).Must(UniqueOptionValidator).WithMessage(localizationService.GetResource("Admin.Catalog.Categories.Fields.PageSizeOptions.ShouldHaveUniqueItems"));
-        }
-
-        private bool UniqueOptionValidator(string input)
-        {
-            if (string.IsNullOrEmpty(input))
-            {
-                return true;
-            }
-            var notValid = input.Split(',').Select(p => p.Trim()).GroupBy(p => p).Any(p => p.Count() > 1);
-            return !notValid;
+            RuleFor(x => x.PageSizeOptions).Must(ValidatorUtilities.PageSizeOptionsValidator).WithMessage(localizationService.GetResource("Admin.Catalog.Categories.Fields.PageSizeOptions.ShouldHaveUniqueItems"));
         }
     }
 }
