+++ src/Presentation/Nop.Web/Administration/Validators/Catalog/ManufacturerValidator.cs
@@ -10,6 +10,8 @@ namespace Nop.Admin.Validators.Catalog
         public ManufacturerValidator(ILocalizationService localizationService)
         {
             RuleFor(x => x.Name).NotEmpty().WithMessage(localizationService.GetResource("Admin.Catalog.Manufacturers.Fields.Name.Required"));
+
+            RuleFor(x => x.PageSizeOptions).Must(ValidatorUtilities.PageSizeOptionsValidator).WithMessage(localizationService.GetResource("Admin.Catalog.Manufacturers.Fields.PageSizeOptions.ShouldHaveUniqueItems"));
         }
     }
 }
