+++ src/Presentation/Nop.Web/Administration/Validators/Vendors/VendorValidator.cs
@@ -13,6 +13,8 @@ namespace Nop.Admin.Validators.Vendors
 
             RuleFor(x => x.Email).NotEmpty().WithMessage(localizationService.GetResource("Admin.Vendors.Fields.Email.Required"));
             RuleFor(x => x.Email).EmailAddress().WithMessage(localizationService.GetResource("Admin.Common.WrongEmail"));
+
+            RuleFor(x => x.PageSizeOptions).Must(ValidatorUtilities.PageSizeOptionsValidator).WithMessage(localizationService.GetResource("Admin.Vendors.Fields.PageSizeOptions.ShouldHaveUniqueItems"));
         }
     }
 }
